requirements.txt
fastapi
uvicorn
apify-client
openai
python-dotenv
pydantic
httpx
